requirements.txt
langchain
faiss-cpu
python-dotenv
PyPDF2
openai
tqdm
docx2txt
